requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0
openai==1.59.7
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.10.6
streamlit==1.41.1
tabulate==0.9.0
python-dotenv==1.0.1